Implement proper checkpointing support; jobs fixes
